% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_site_mean_sd_plotter.R
\name{all_site_mean_sd_plotter}
\alias{all_site_mean_sd_plotter}
\title{Daily plotter for all sites for a given variable}
\usage{
all_site_mean_sd_plotter(
  list_df,
  y_label,
  y_col = "mean_FCH4",
  sd_col = "sd_FCH4",
  show_x_label = FALSE,
  ignore_NA = TRUE,
  GPP = FALSE,
  Reco = FALSE,
  NEE = FALSE
)
}
\arguments{
\item{list_df}{a list of dataframes having the same structure and able to be bound by row. When this comes from the saved aggregated tables, they are just the _day element of the list.}

\item{y_label}{label of the y axis. it can also be a bquote if superscripts or underscripts are needed.}

\item{y_col}{string, name of the variable to be represented on the y axis}

\item{sd_col}{string, name of the variable to use for the ribbon}

\item{show_x_label}{logical, TRUE to show x-axis text and title, FALSE to hider x-axis text and title.#'}
}
\value{
a ggplot object
}
\description{
Daily plotter for all sites for a given variable
}
