% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomet_aggregator.R
\name{biomet_aggregator}
\alias{biomet_aggregator}
\title{Biomet variables aggregator, from halfhourly to either hourly, daily, weekly, monthly yearly and growing_season timescales}
\usage{
biomet_aggregator(
  df = gapfilling_list_hals$site_df,
  datetime = "datetime",
  biomet_vars = c("Ta_f", "Ts_f", "WTD_f", "PARin_f", "SNOWd"),
  aggr_fun = c("mean", "mean", "mean", "sum", "sum"),
  aggr_var = c("hour", "day", "week", "month", "year", "growing_season"),
  month_start_growing_season = 5,
  month_end_growing_season = 10
)
}
\arguments{
\item{df}{dataframe containing the variables to aggregate and the timestamp column}

\item{datetime}{timestamp column name already in a POSIXct format and UTC timezone to avoid summertime problems}

\item{biomet_vars}{Vector of Biomet variables to aggregate}

\item{aggr_fun}{Vector of aggregation functions to apply to each biomet variable. This should be the same lendth and same order as biomet_vars}

\item{aggr_var}{Type ofaggregation, either of "hour","day","week","month","year","growing_season"}

\item{month_start_growing_season}{When aggr_var is "growing_seaon", this is the month considered as the start of the growing season}

\item{month_end_growing_season}{When aggr_var is "growing_seaon", this is the month considered as the end of the growing season}
}
\value{
a summarized dataframe, containing the aggregation timescale variables and the aggregated values
}
\description{
Biomet variables aggregator, from halfhourly to either hourly, daily, weekly, monthly yearly and growing_season timescales
}
