% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_air_temperature_from_halfhour.R
\name{daily_air_temperature_from_halfhour}
\alias{daily_air_temperature_from_halfhour}
\title{Aggregates halfhourly air temperature data to average daily temperatures. Useful for defining the growing season even before aggregating biomet and fluxes}
\usage{
daily_air_temperature_from_halfhour(df, time_col, Ta_col)
}
\arguments{
\item{df}{a dataframe containing a datetime column as POSIXct object, and an air temperature column in °C}

\item{time_col}{the name of the datetime column}

\item{Ta_col}{the name of the air temperature column}
}
\value{
a dataframe containing daily values of average air temperature with 2 columns: date, and Ta.
}
\description{
Aggregates halfhourly air temperature data to average daily temperatures. Useful for defining the growing season even before aggregating biomet and fluxes
}
