% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_temperature_amplitude.R
\name{daily_temperature_amplitude}
\alias{daily_temperature_amplitude}
\title{Computes temperature (soil temperature, but can also be air temperature) daily amplitudes from halfhourly data}
\usage{
daily_temperature_amplitude(df, time_col, T_col)
}
\arguments{
\item{df}{a dataframe containing a datetime column as POSIXct object, and temperature column in °C}

\item{time_col}{the name of the datetime column}

\item{T_col}{the name of the temperature column}
}
\value{
a dataframe containing daily values of daily temperature amplitude, and the mean soil temperature#' @export
}
\description{
Computes temperature (soil temperature, but can also be air temperature) daily amplitudes from halfhourly data
}
