% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dte_formatter.R
\name{dte_formatter}
\alias{dte_formatter}
\title{Date formatter for use in other ggplot related functions}
\usage{
dte_formatter(x, mode = "month_only")
}
\arguments{
\item{x}{: a date}

\item{mode}{: one of "month_only" or "month_year"}
}
\value{
formatted dates
}
\description{
Date formatter for use in other ggplot related functions
}
