% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_fluxes_REddyProc.R
\name{formatting_fluxes_REddyProc}
\alias{formatting_fluxes_REddyProc}
\title{Preparation of flux data for REddyProc gapfilling}
\usage{
formatting_fluxes_REddyProc(
  df,
  datetime = "datetime",
  flux_col = "co2_flux_final",
  FLUX = c("NEE", "H2O"),
  rg_col = "SWin_f",
  tair_col = "Ta_f",
  tsoil_col = "Ts_f",
  rh_col = "RH_f",
  vpd_col = "VPD_f",
  ustar_col = "u*",
  saving_path,
  filename = NULL,
  use_tsoil_col = TRUE
)
}
\arguments{
\item{df}{Dataframe containing the flux column, timestamp and all the biomet variables.
If following the standard workflow, use the output from the quality check}

\item{datetime}{timestamp column name already as a POSIXct object}

\item{flux_col}{flux column, already quality checked/controlled}

\item{FLUX}{flux column in the final output: NEE for CO2 flux OR H2O for water flux}

\item{rg_col}{global radiation column name or other variable used in its place (e.g. SWin)}

\item{tair_col}{air temperature column name}

\item{tsoil_col}{soil temperature column name}

\item{rh_col}{relative humidity column name}

\item{vpd_col}{vapor pressure deficit column name}

\item{ustar_col}{u* column name}

\item{saving_path}{saving folder path where the formatted file will be saved}

\item{filename}{WITHOUT extension. name to be appended to saving_path to create full path name. If NULL, a default name will be written}

\item{use_tsoil_col}{Boolean variable, default is TRUE, set to FALSE TO IGNORE TSOIL}
}
\value{
No value returned, but a file saved to disk
}
\description{
Preparation of flux data for REddyProc gapfilling
}
