% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfill_between_sites.R
\name{gapfill_between_sites}
\alias{gapfill_between_sites}
\title{Check correlations and Gapfill environmental variables between sites}
\usage{
gapfill_between_sites(
  datetime = "datetime",
  df_to_fill,
  ind_df1,
  ind_df2 = NULL,
  ind_df3 = NULL,
  col_to_fill,
  ind_col1 = col_to_fill,
  ind_col2 = NULL
)
}
\arguments{
\item{datetime}{name of datetime, POSIXct object column in each table. This name should be the same between sites}

\item{df_to_fill}{dataframe containing the column to gapfill}

\item{ind_df1}{first dataframe containing the column or columns to use in the regression}

\item{ind_df2}{second dataframe containing the  column or columns to use in the regression}

\item{ind_df3}{third dataframe containing the  column or columns  to use in the regression}

\item{col_to_fill}{name of the column in df_to_fill, with missing data to fill}

\item{ind_col1}{default, same name as col_to_fill; is the first independent column name from either of ind_df1, ind_df2, ind_df3, to be used to fill gaps in col_to_fill. Regression will be used and one dataframe}

\item{ind_col2}{second independent column, to be used in case first column is absent, or if multiple regression will be tried}
}
\value{
the same dataframe with the column gapfilled
}
\description{
Check correlations and Gapfill environmental variables between sites
}
