% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_timeseries_linear.R
\name{gf_timeseries_linear}
\alias{gf_timeseries_linear}
\title{Function for gap-filling time series using linear interpolation based on dates.}
\usage{
gf_timeseries_linear(dates, values)
}
\arguments{
\item{dates}{vector of dates without missing values}

\item{values}{vector of values with missing values to gapfill}
}
\value{
a vector of the same length as dates and values, where gaps are filled. outside gaps are not filled
}
\description{
Function for gap-filling time series using linear interpolation based on dates.
}
\examples{
gf_timeseries_linear(as.Date(c("2023-07-01", "2023-07-02", "2023-07-03", "2023-07-07", "2023-07-08", "2023-07-10", "2023-07-11", "2023-07-15", "2023-07-16")),
c(NA,34, NA, NA, 5, NA, NA, 15, 9))
}
