% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growing_season_definer.R
\name{growing_season_definer}
\alias{growing_season_definer}
\title{Growing season definition from date or timestamp and air temperature}
\usage{
growing_season_definer(
  df,
  time_col,
  Ta_col,
  Ta_day_threshold = 5,
  number_consec_days = 5
)
}
\arguments{
\item{df}{dataframe containing a timestamp or date, and air temperature}

\item{time_col}{name of the timestamp column. It can be a Date column, or a POSIXct object}

\item{Ta_col}{name of the air temperature (in °C) column}

\item{Ta_day_threshold}{threshold of air temperature above or below which the growing season starts or ends}

\item{number_consec_days}{number of consecutive days when Ta is above or below Ta_day_threshold for start/stop counting the growing season}
}
\value{
a dataframe with 3 columns: the year, the date of start of the growing season, the date of end of the growing season
}
\description{
Growing season definition from date or timestamp and air temperature
}
