% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.R
\name{laplace}
\alias{laplace}
\title{Laplace distribution}
\usage{
laplace(x, m = 0, t = 1)
}
\arguments{
\item{x}{x value, a single value or vector for which the laplace y value will be calculated}

\item{m}{position parameter}

\item{t}{scale parameter}
}
\value{
y, same dimension as x, the laplace "transformed" value or vector
}
\description{
Laplace distribution
}
\examples{
laplace(2,0,1)
}
