% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laprnd.R
\name{laprnd}
\alias{laprnd}
\title{generate i.i.d.(Independent and identically distributed) laplacian random number drawn from laplacian distribution}
\usage{
laprnd(mu, sigma)
}
\arguments{
\item{mu}{mean of the double exponential (laplacian) distribution}

\item{sigma}{standard deviation of the double exponential (laplacian) distribution}
}
\value{
laplace random noise
}
\description{
generate i.i.d.(Independent and identically distributed) laplacian random number drawn from laplacian distribution
}
\examples{
laprnd(0,1)
}
