% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_2_EC_date.R
\name{merge_2_EC_date}
\alias{merge_2_EC_date}
\title{Merge 2 Eddypro outputs, and choose to save the output to disk, but always return the merged dataframe. In this "date" version, you state the datetime where to merge, to remove overlaps}
\usage{
merge_2_EC_date(
  path_EC1,
  path_EC2,
  datetime_merge,
  date_format_EC1 = "\%Y-\%m-\%d",
  date_format_EC2 = "\%Y-\%m-\%d",
  check_write_df = FALSE,
  writing_path = NULL
)
}
\arguments{
\item{path_EC1}{path to the first eddypro full output file}

\item{path_EC2}{path to the second eddypro full output file}

\item{datetime_merge}{timestamp as string, format "%YYYY-%mm-%dd %HH:%MM:%SS" at which to remove rows after -for df1- and rows before -for df2- . Note that the specified timestamp itself will be included in df2 and excluded from df1}

\item{date_format_EC1}{date format of the date column in the first eddypro full output. If the eddypro output has not been altered, the date format is "\%Y-\%m-\%d", the default. If it has been opened in excel for instance, the date format is likely "\%d/\%m/\%Y"}

\item{date_format_EC2}{date format of the date column in the second eddypro full output. If the eddypro output has not been altered, the date format is "\%Y-\%m-\%d", the default. If it has been opened in excel for instance, the date format is likely "\%d/\%m/\%Y"}

\item{check_write_df}{logical, TRUE if you want to save the output to disk, in which case you must also provide writing_path}

\item{writing_path}{string, folder to which the merged file will be saved. the file name "merged.csv" will be appended automatically to the folder path. The folder path should not end with a slash.}
}
\value{
the merged dataframe
}
\description{
Merge 2 Eddypro outputs, and choose to save the output to disk, but always return the merged dataframe. In this "date" version, you state the datetime where to merge, to remove overlaps
}
