% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmolH2Om2s_to_mm_30min.R
\name{mmolH2Om2s_to_mm_30min}
\alias{mmolH2Om2s_to_mm_30min}
\title{mmolH2O.m-2.s-1 to mm.30min-1 to mmH2O.30min-1 Converter (useful to have water fluxes that can be summed up directly to aggregate over day, month, year after gapfilling)}
\usage{
mmolH2Om2s_to_mm_30min(
  h2o_flux,
  M_h2O = 18,
  mmol_mol = 10^-3,
  min30_s = 30 * 60,
  mL_m3 = 10^-6,
  m_mm = 10^3
)
}
\arguments{
\item{h2o_flux}{single value or vector representing water fluxes in mmolH2O.m-2.s-1}

\item{M_h2O}{molar mass of water}

\item{mmol_mol}{constant for converting from mmol to mol}

\item{min30_s}{conversion factor of 30 min to seconds}

\item{mL_m3}{constant for converting mililiters to cubic meter}

\item{m_mm}{constant for converting from meters to millimeters}
}
\value{
water fluxes in mm.30min-1 to mmH2O.30min-1, the same length as h2o_flux
}
\description{
mmolH2O.m-2.s-1 to mm.30min-1 to mmH2O.30min-1 Converter (useful to have water fluxes that can be summed up directly to aggregate over day, month, year after gapfilling)
}
\examples{
mmolH2Om2s_to_mm_30min(1)
}
