% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo_sim_gf.R
\name{montecarlo_sim_gf}
\alias{montecarlo_sim_gf}
\title{Random Forest Gapfilling of montecarlo simulated datasets}
\usage{
montecarlo_sim_gf(mc_sim_path, mtry, preds, flux_col, mc_sim_gf_path)
}
\arguments{
\item{mc_sim_path}{path to the montecarlo simulation files. This folder should only contain those files and nothing else.}

\item{mtry}{the best mtry found during training in the real gapfilling. This should be fetched from the previously saved model.}

\item{preds}{same predictors as used in the real gapfilling}

\item{flux_col}{column holding the fluxes to be gapfilled}

\item{mc_sim_gf_path}{saving folder, where the gpafilled output should be saved}
}
\value{
this function will return nothing, but files should be saved in a specified folder
}
\description{
Random Forest Gapfilling of montecarlo simulated datasets
}
