% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo_sim_gf_CO2_xgb.R
\name{montecarlo_sim_gf_CO2_xgb}
\alias{montecarlo_sim_gf_CO2_xgb}
\title{Random Forest Gapfilling of montecarlo simulated datasets}
\usage{
montecarlo_sim_gf_CO2_xgb(
  mc_sim_path,
  best_tune,
  preds,
  flux_col,
  mc_sim_gf_path,
  datetime = "datetime"
)
}
\arguments{
\item{mc_sim_path}{path to the montecarlo simulation files. This folder should only contain those files and nothing else.}

\item{best_tune}{the best tuning grid obtained from the Caret training, used for the gapfilling}

\item{preds}{same predictors as used in the real gapfilling}

\item{flux_col}{column holding the fluxes to be gapfilled}

\item{mc_sim_gf_path}{saving folder, where the gpafilled output should be saved}

\item{datetime}{datetime column name as a POSIXct object}
}
\value{
this function will return nothing, but files should be saved in a specified folder
}
\description{
Random Forest Gapfilling of montecarlo simulated datasets
}
