% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo_sim_noCH4.R
\name{montecarlo_sim_noCH4}
\alias{montecarlo_sim_noCH4}
\title{Generating n tables with each having a random laplacian noise added to the original fluxes, so called montecarlo simulation (For ReddyProc)}
\usage{
montecarlo_sim_noCH4(
  path_to_file,
  flux_col = c("NEE", "H2O"),
  flux_pred_col = c("NEE_fall", "H2O_fall"),
  n = 100,
  saving_folder
)
}
\arguments{
\item{path_to_file}{path to the gapfilled flux dataframe (Reddyproc gapfilled)}

\item{flux_col}{original flux column: 'NEE' for CO2 flux OR 'H2O' for water flux}

\item{flux_pred_col}{gapfilled flux column: 'NEE_fall' for CO2 flux OR 'H2O_fall' for water flux}

\item{n}{Number of montecarlo iterations}

\item{saving_folder}{path to the folder where the simulations will be saved}
}
\value{
No value returned, but a file saved to disk
}
\description{
Generating n tables with each having a random laplacian noise added to the original fluxes, so called montecarlo simulation (For ReddyProc)
}
