% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo_sim_noCH4_gf.R
\name{montecarlo_sim_noCH4_gf}
\alias{montecarlo_sim_noCH4_gf}
\title{REddyProc gapfilling of all}
\usage{
montecarlo_sim_noCH4_gf(
  mc_sim_path,
  mc_sim_gf_path,
  flux_col = c("NEE", "H2O"),
  longi = 19.556646,
  lati = 64.18198,
  timez = 1
)
}
\arguments{
\item{mc_sim_path}{path to the montecarlo simulated files}

\item{mc_sim_gf_path}{path where the gapfilled simulations will be saved}

\item{flux_col}{flux column, either of "NEE" or "H2O"}

\item{longi}{longitude of the site}

\item{lati}{latitude of the site}

\item{timez}{timezone as a signed integer (e.g. 1 for UTC+1)}
}
\value{
no value returned, but files are gapfilled and saved to disk
}
\description{
REddyProc gapfilling of all
}
