% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo_sim_old.R
\name{montecarlo_sim_old}
\alias{montecarlo_sim_old}
\title{Generating n tables with each having a random laplacian noise added to the original fluxes, so called montecarlo simulation}
\usage{
montecarlo_sim_old(
  df_gf,
  flux_col = "ch4_flux_final",
  flux_pred_col = "predicted",
  datetime = "datetime",
  preds,
  n,
  saving_folder
)
}
\arguments{
\item{df_gf}{gapfilled flux dataframe. This would be retrieved for instance from the rf_gapfiller function}

\item{flux_col}{column name of the non-gapilled fluxes}

\item{flux_pred_col}{column name of the predicted fluxes (predictions of the non-gap values)}

\item{datetime}{datetime column}

\item{preds}{same predictors used for gapfilling}

\item{n}{number of montecarlo simulations to run}

\item{saving_folder}{folder where the simulations will be saved}
}
\value{
this function will return nothing, but files should be saved in a specified folder
}
\description{
Generating n tables with each having a random laplacian noise added to the original fluxes, so called montecarlo simulation
}
