% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter_3D.R
\name{plotter_3D}
\alias{plotter_3D}
\title{A wrapper for plotly's 3D scatterplots.}
\usage{
plotter_3D(df, x_var, y_var, z_var, col_var)
}
\arguments{
\item{df}{dataframe of daily sum of fluxes and average of environmental variables}

\item{x_var}{variable on x axis}

\item{y_var}{variabke on y axis}

\item{z_var}{variable on z axis}

\item{col_var}{variable for coloring the points}
}
\value{
A 3D graph
}
\description{
A wrapper for plotly's 3D scatterplots.
}
