% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter_daily.R
\name{plotter_daily}
\alias{plotter_daily}
\title{Daily plotter for all sites for a given variable}
\usage{
plotter_daily(
  list_df,
  y_label,
  y_col = "FCH4_sum",
  show_x_label = FALSE,
  ignore_NA = TRUE
)
}
\arguments{
\item{list_df}{a list of dataframes having the same structure and able to be bound by row. When this comes from the saved aggregated tables, they are just the _day element of the list.}

\item{y_label}{label of the y axis. it can also be a bquote if superscripts or underscripts are needed.}

\item{y_col}{string, name of the variable to be represented on the y axis.}

\item{show_x_label}{logical, TRUE to show x-axis text and title, FALSE to hider x-axis text and title.}

\item{ignore_NA}{logical. If TRUE, segments without NA will be plotted separately to avoid interpolation of geom_smooth where there are missing values}
}
\value{
a ggplot object
}
\description{
Daily plotter for all sites for a given variable
}
