% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter_daily_avg.R
\name{plotter_daily_avg}
\alias{plotter_daily_avg}
\title{Daily plotter for all sites for a given variable}
\usage{
plotter_daily_avg(
  list_df,
  y_label,
  y_col = "FCH4_sum",
  show_x_label = FALSE,
  ignore_NA = TRUE,
  GPP = FALSE,
  Reco = FALSE,
  NEE = FALSE,
  yintercept_line = FALSE,
  yintercept_value = 0,
  scale_x = "month_only",
  rotate_x_label = FALSE,
  conf_int = FALSE,
  window_size = 14,
  sd_col = NULL,
  y_col_ori = NULL,
  different_y_line_point = FALSE
)
}
\arguments{
\item{list_df}{a list of dataframes having the same structure and able to be bound by row. When this comes from the saved aggregated tables, they are just the _day element of the list.}

\item{y_label}{label of the y axis. it can also be a bquote if superscripts or underscripts are needed.}

\item{y_col}{string, name of the variable to be represented on the y axis.}

\item{show_x_label}{logical, TRUE to show x-axis text and title, FALSE to hider x-axis text and title.}

\item{ignore_NA}{logical. If TRUE, segments without NA will be plotted separately to avoid interpolation of geom_smooth where there are missing values}

\item{GPP}{logical. If TRUE, the graph us then tailored for GPP. default is FALSE}

\item{Reco}{logical. If TRUE, the graph us then tailored for Reco. default is FALSE}

\item{NEE}{logical. If TRUE, the graph us then tailored for NEE. default is FALSE}

\item{yintercept_line}{logical. If TRUE, a horizontal line will be added. default is FALSE}

\item{yintercept_value}{positive integer. specifies the y value at which the horizonal line will be drawn. default is 0}

\item{scale_x}{One of "month_only or "month_year", default is month_only}

\item{rotate_x_label}{Logical. default, FALSE}

\item{conf_int}{Logical. Display a ribbon confidence interval around the smoothed lines. default, FALSE.}

\item{window_size}{window size for centered rolling average, default is 14 days}

\item{sd_col}{column to use for error when conf_int is TRUE and we want to plot a ribbon. If not provided and column is NULL, a rolling sd will be computed with the same window_size as for the moving average line...}

\item{y_col_ori}{second Y column to use for points. useful for instance when y has been na.approximated to gapfill missing data and you still want to plot points with the original, and rollmeans with the interpolated
Remember to set different_y_line_point to TRUE for it to work.}

\item{different_y_line_point}{logical, default FALSE. set to TRUE along side y_col_ori to use y_col_ori to plot points}
}
\value{
a ggplot object
}
\description{
Daily plotter for all sites for a given variable
}
