% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter_monthly_anomalies.R
\name{plotter_monthly_anomalies}
\alias{plotter_monthly_anomalies}
\title{Plotting monthly anomalies average from yearly or growing-season}
\usage{
plotter_monthly_anomalies(
  list_df,
  x_label = "Monthly GPP anomaly (gC. m-2. month-1)",
  y_label = "Monthly CH4 flux anomaly (g. m-2. month-1)",
  x_col = "GPP_sum",
  y_col = "FCH4_sum",
  full_periods_only = TRUE,
  show_x_label = TRUE,
  show_y_label = FALSE,
  add_geom_smooth = TRUE
)
}
\arguments{
\item{list_df}{a list of dataframes having the same structure and able to be bound by row. When this comes from the saved aggregated tables, they are just the _month element of the list.}

\item{x_label}{Label to write on the x axis}

\item{y_label}{Label to write on the y axis}

\item{x_col}{the x column name}

\item{y_col}{the y column name}

\item{full_periods_only}{logical. If TRUE, full years only will be used}

\item{show_x_label}{logical. Show or not show x label}

\item{show_y_label}{logical. Show or not show x label}

\item{add_geom_smooth}{logical. Show or not a linear fit}
}
\value{
a ggplot object
}
\description{
Plotting monthly anomalies average from yearly or growing-season
}
