% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter_monthly_anomalies_2.R
\name{plotter_monthly_anomalies_2}
\alias{plotter_monthly_anomalies_2}
\title{Plotting monthly anomalies average of every month over the years}
\usage{
plotter_monthly_anomalies_2(
  list_df,
  period_start = "2020-04-01",
  period_end = "2022-03-31",
  x_label = "Monthly GPP anomaly (gC. m-2. month-1)",
  y_label = "Monthly CH4 flux anomaly (g. m-2. month-1)",
  x_col = "GPP_sum",
  y_col = "FCH4_sum",
  full_periods_only = TRUE,
  show_x_label = TRUE,
  show_y_label = FALSE,
  add_geom_smooth = TRUE,
  color_grey = TRUE,
  add_rsq = TRUE,
  growing_season = FALSE,
  levels_sites = c("Stortjarn", "Halsingfors", "Halmyran", "Degero")
)
}
\arguments{
\item{list_df}{a list of dataframes having the same structure and able to be bound by row. When this comes from the saved aggregated tables, they are just the _month element of the list.}

\item{period_start}{start date in the format YYYY-mm-dd}

\item{period_end}{End date in the format YYYY-mm-dd}

\item{x_label}{Label to write on the x axis}

\item{y_label}{Label to write on the y axis}

\item{x_col}{the x column name}

\item{y_col}{the y column name}

\item{full_periods_only}{logical. If TRUE, full years only will be used}

\item{show_x_label}{logical. Show or not show x label}

\item{show_y_label}{logical. Show or not show x label}

\item{add_geom_smooth}{logical. Show or not a linear fit}

\item{color_grey}{logical. color in grey or not}

\item{add_rsq}{whether or not to add r-squared}

\item{growing_season}{logical, whether or not to limit the dataset to the growing season (May to October)}

\item{levels_sites}{a vector showing the order of the site names.}
}
\value{
a ggplot object
}
\description{
Plotting monthly anomalies average of every month over the years
}
