% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{quality_control}
\alias{quality_control}
\title{Title}
\usage{
quality_control(
  fluxes_meteo,
  mauden_flag = c(1, 2),
  hard_flags = TRUE,
  limit_lower = -0.1,
  limit_upper = 1,
  RSSI = FALSE,
  RSSI_col = "diag_77_mean",
  ustar = TRUE,
  Temp_col = "Ts",
  datetime_start = NULL,
  datetime_end = NULL
)
}
\arguments{
\item{fluxes_meteo}{dataframe containing fluxes and biomet data}

\item{mauden_flag}{vector, default c(1,2) for qc values to flag out. Recommended practice is to keep 0 as good data, and flag out 2 only if dealing with annual budgets, and flag both 1 and 2 if looking at high frequency data}

\item{hard_flags}{boolean telling whether or not to apply a hard flag}

\item{limit_lower}{lower limit for the hard flags after Holl et al. 2020}

\item{limit_upper}{upper limit for the hard flags after Holl et al. 2020}

\item{RSSI}{boolean telling whether or not to do a RSSI filtering. Default is TRUE}

\item{RSSI_col}{name of the column that contains RSSI values.}

\item{ustar}{boolean telling whether or not to do u* filtering. Default is TRUE}

\item{Temp_col}{name of the temperature column to be used for u_star threshold calculation}

\item{datetime_start}{starting datetime as a posixct object (format \%Y-\%m-\%d \%H:\%M:\%S") for creating a complete dataset to ensure no missing values exist in the final dataset. If not provided (default NULL), the first/minimum datetime will be used.}

\item{datetime_end}{Ending datetime as a posixct object (format \%Y-\%m-\%d \%H:\%M:\%S") for creating a complete dataset to ensure no missing values exist in the final dataset. If not provided (default NULL), the last/maximum datetime will be used.}
}
\value{
a dataframe having some additional columns, with the flux column suffixed by "_final" being the quality checked flux values
}
\description{
Title
}
