% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddyproc_gapfiller_old.R
\name{reddyproc_gapfiller_old}
\alias{reddyproc_gapfiller_old}
\title{A wrapper for ReddyProc gapfiller for gapfilling NEE or H2O fluxes,and partitionning NEE. This old form has a problem. Due to using sMRFluxPartition instead of sMRFluxPartitionUStarScens, Reco was not as variable as it should have been.}
\usage{
reddyproc_gapfiller_old(
  formatted_file_path,
  saving_folder,
  file_name = NULL,
  FLUX = c("NEE", "H2O"),
  longitude = 19.556646,
  latitude = 64.18198,
  timezone = 1,
  gapfill_flux = TRUE
)
}
\arguments{
\item{formatted_file_path}{path to the file ready for ReddyProc gapfilling. Typically the output from the "formatting_fluxes_REddyProc" function of the PostEddyPro package}

\item{saving_folder}{path to the folder where the gapfilled fluxes will be saved}

\item{file_name}{fine name with extension (only base name with extension, e.g. fluxes.txt)}

\item{FLUX}{flux column in the final output: NEE for CO2 flux OR H2O for water flux}

\item{longitude}{longitude of the site}

\item{latitude}{latitude of the site}

\item{timezone}{timezone as a signed integer (e.g. 1 for UTC+1)}

\item{gapfill_flux}{boolean. Default is TRUE. Set to false if gapfilling of NEE has been performed before and do only partitioning.}
}
\value{
No value returned, but a file saved to disk
}
\description{
A wrapper for ReddyProc gapfiller for gapfilling NEE or H2O fluxes,and partitionning NEE. This old form has a problem. Due to using sMRFluxPartition instead of sMRFluxPartitionUStarScens, Reco was not as variable as it should have been.
}
