% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_gapfiller.R
\name{rf_gapfiller}
\alias{rf_gapfiller}
\title{A randomForest gapfiller, with an integrated training-testing from caret}
\usage{
rf_gapfiller(
  site_df,
  datetime = "datetime",
  flux_col = "ch4_flux_final",
  preds,
  max_mtry = ceiling(max(sqrt(length(preds)), length(preds)/3)),
  sitename
)
}
\arguments{
\item{site_df}{dataframe containing all a POSIXct timestamp, the flux data and predictor columns (preds argument).}

\item{datetime}{name of the timestamp (datetime) column already in POSIXct format}

\item{flux_col}{flux column to be gapfilled, this function was developped for methane fluxes, but can be used for other fluxes as long as a good set of predictors are provided}

\item{preds}{vector of predictors that have been gapfilled in advance}

\item{max_mtry}{maximum mtry parameter to be used in the random forest. it is calculated here as the maximum of sqrt of the number of predictors, and the 1/3rd of the number of predictors}

\item{sitename}{sitename just for labelling the plots}
}
\value{
large list of elements. tuningmodel: finalmodel (the model used for gapfilling), site_df (the final data.frame containing both the input dataframe and the gapfilled column), r_sq_cv (crossvalidated R-squared)
rmse_cv (crossvalidated rmse), pred_meas (graph showing predicted vs. measured values), gf_meas_time (a graph showing both measured and gapfilled values with time)
}
\description{
A randomForest gapfiller, with an integrated training-testing from caret
}
