% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssi_filter.R
\name{rssi_filter}
\alias{rssi_filter}
\title{Filters for low residual signal strength indicator aka RSSI}
\usage{
rssi_filter(fluxes_meteo, gas, RSSI_col)
}
\arguments{
\item{fluxes_meteo}{dataframe containing both fluxes and soil/air temperature data}

\item{gas}{gas name as denoted in Eddypro (e.g. "ch4" or "co2",includes also non gas like LE)}

\item{RSSI_col}{name of the column containing RSSI values}
}
\value{
a list containing the RSSI threshold decided by the user based on the graph (the visual inflection point), and a vector of flags (1 means low RSSI, and 0 means good RSSI) the same length as the number of rows of the input data.frame
}
\description{
Filters for low residual signal strength indicator aka RSSI
}
