% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_calculators.R
\name{temporal_calculators}
\alias{temporal_calculators}
\title{Caculates and appends sine, cosine and time delta to a dataframe}
\usage{
temporal_calculators(df, datetime = "datetime")
}
\arguments{
\item{df}{dataframe containing the timestamp column}

\item{datetime}{timestamp column name, which should already be a POSIXct object}
}
\value{
a dataframe to which the sine, cosine and delta (decimal timestamp from beginning of the year)
}
\description{
Caculates and appends sine, cosine and time delta to a dataframe
}
\examples{
temporal_calculators(data.frame(datetime=as.POSIXct(c("2020-01-01 00:30:00", "2020-01-01 01:00:00", "2020-01-01 01:30:00", "2020-01-01 02:00:00"))))
}
