% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umolCH4m2s_to_mgCH4m2_30min.R
\name{umolCH4m2s_to_mgCH4m2_30min}
\alias{umolCH4m2s_to_mgCH4m2_30min}
\title{umolCH4.m-2.s-2 to mgCH4.m-2.30min-1 Converter (useful to have methane fluxes that can be summed up directly to aggregate over day, month, year after gapfilling)}
\usage{
umolCH4m2s_to_mgCH4m2_30min(
  ch4_flux,
  umol_mol = 10^-6,
  M_CH4 = 16.04,
  g_mg = 10^3,
  min30_s = 30 * 60
)
}
\arguments{
\item{ch4_flux}{methane flux in umolCH4.m-2.s-2, a single value or a vector}

\item{umol_mol}{constant factor for converting umol to mol}

\item{M_CH4}{molar mass of methane}

\item{g_mg}{constant factor for converting g to mg}

\item{min30_s}{conversion factor of 30 min to seconds}
}
\value{
methane flux in mgCH4.m-2.30min-1, the same size as ch4_flux
}
\description{
umolCH4.m-2.s-2 to mgCH4.m-2.30min-1 Converter (useful to have methane fluxes that can be summed up directly to aggregate over day, month, year after gapfilling)
}
\examples{
umolCH4m2s_to_mgCH4m2_30min(0.2)
}
