% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umolCO2m2s_to_gCm2_30min.R
\name{umolCO2m2s_to_gCm2_30min}
\alias{umolCO2m2s_to_gCm2_30min}
\title{umolCO2.m-2.s-2 to gC.m-2.30min-1 Converter (useful to have CO2 fluxes that can be summed up directly to aggregate over day, month, year after gapfilling)}
\usage{
umolCO2m2s_to_gCm2_30min(
  co2_flux,
  umol_mol = 1e-06,
  M_C = 12.011,
  min30_s = 30 * 60
)
}
\arguments{
\item{co2_flux}{CO2 flux in umolCO2.m-2.s-2, a single value or a vector}

\item{umol_mol}{constant factor for converting umol to mol}

\item{M_C}{molar mass of carbon}

\item{min30_s}{conversion factor of 30 min to seconds}
}
\value{
CO2 flux in gC.m-2.30min-1, the same size as co2_flux
}
\description{
umolCO2.m-2.s-2 to gC.m-2.30min-1 Converter (useful to have CO2 fluxes that can be summed up directly to aggregate over day, month, year after gapfilling)
}
\examples{
umolCO2m2s_to_gCm2_30min(1)
}
