% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustar_filter.R
\name{ustar_filter}
\alias{ustar_filter}
\title{A function for change point detection of the u* threshold for low turbulence}
\usage{
ustar_filter(
  fluxes_meteo,
  gas,
  Temp_col = "Ts",
  number_Temp_classes = 6,
  number_ustar_classes = 20,
  ustar_maximum_limit = 0.8
)
}
\arguments{
\item{fluxes_meteo}{dataframe containing both fluxes and soil/air temperature data}

\item{gas}{gas name as denoted in Eddypro (e.g. "ch4" or "co2",includes also non gas like LE)}

\item{Temp_col}{temperature column}

\item{number_Temp_classes}{number of bin of soil/air temperature to be used, default is 6}

\item{number_ustar_classes}{number of bin of ustar to be used, default is 20}

\item{ustar_maximum_limit}{an upper limit to exclude for ustar. This argument is particularly useful when abnormally high values are contained in ustar...}
}
\value{
a list containing the ustar threshold based on the change point detection method, and a vector of flags (1 means low turbulence or ustar below thresold and 0 means enough turbulence) the same length as the number of rows of the input data.frame
}
\description{
A function for change point detection of the u* threshold for low turbulence
}
