% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_gapfiller.R
\name{xgboost_gapfiller}
\alias{xgboost_gapfiller}
\title{An XGBoost gapfiller, with an integrated tuning-training-testing from caret}
\usage{
xgboost_gapfiller(
  site_df,
  datetime = "datetime",
  flux_col = "co2_flux",
  preds,
  sitename
)
}
\arguments{
\item{site_df}{dataframe containing all a POSIXct timestamp, the flux data and predictor columns (preds argument).}

\item{datetime}{name of the timestamp (datetime) column already in POSIXct format}

\item{flux_col}{flux column to be gapfilled, this function was developped for methane fluxes, but can be used for other fluxes as long as a good set of predictors are provided}

\item{preds}{vector of predictors that have been gapfilled in advance}

\item{sitename}{sitename just for labelling the plots}
}
\value{
large list of elements. tuningmodel: finalmodel (the model used for gapfilling), site_df (the final data.frame containing both the input dataframe and the gapfilled column), r_sq_cv (crossvalidated R-squared)
rmse_cv (crossvalidated rmse), pred_meas (graph showing predicted vs. measured values), gf_meas_time (a graph showing both measured and gapfilled values with time)
}
\description{
An XGBoost gapfiller, with an integrated tuning-training-testing from caret
}
