% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomet_aggregator.R
\name{biomet_aggregator}
\alias{biomet_aggregator}
\title{Biomet variables aggregator, from halfhourly to either hourly, daily, weekly, monthly yearly and growing_season timescales}
\usage{
biomet_aggregator(
  aggr_var = c("hour", "day", "week", "month", "year", "growing_season"),
  df = gapfilling_list_hals$site_df,
  datetime = "datetime",
  aggr_fun = c("mean", "mean", "mean", "sum", "mean"),
  biomet_vars = c("Ta_f", "Ts_f", "WTD_f", "PARin_f", "SNOWd"),
  month_start_growing_season = 5,
  month_end_growing_season = 10,
  na.rm_val = TRUE,
  growing_season_definition = "fixed_months",
  df_Ta = NULL,
  df_Ts = NULL,
  number_consec_days = 7
)
}
\arguments{
\item{aggr_var}{Type of aggregation, either of "hour","day","week","month","year","growing_season"}

\item{df}{dataframe containing the variables to aggregate and the timestamp column}

\item{datetime}{timestamp column name already in a POSIXct format and UTC timezone to avoid summertime problems}

\item{aggr_fun}{Vector of aggregation functions to apply to each biomet variable. This should be the same lendth and same order as biomet_vars}

\item{biomet_vars}{Vector of Biomet variables to aggregate}

\item{month_start_growing_season}{When aggr_var is "growing_seaon", this is the month considered as the start of the growing season}

\item{month_end_growing_season}{When aggr_var is "growing_seaon", this is the month considered as the end of the growing season}

\item{na.rm_val}{logical, TRUE if you want NAs to be ignored. Note that this means that sums may be calculated even when there are missing values, which is misleading.}

\item{growing_season_definition}{The way growing season is definied. "fixed_months" (default) will use the months between month_start_growing_season and month_end_growing_season.
"meteorological" will use air temperature data to calculate the growing season start and end dates. This option requires to provide a dataframe containing date and air temperature data.}

\item{df_Ta}{a dataframe containing 2 columns: a column named "date" containing dates, and a column named "Ta" containing daily average air temperatures}

\item{df_Ts}{a dataframe containing 2 columns: a PosiXct column with timestamps in the format "%Y-%m-%d %H:%M:%S" and a column representing soil temperatures (e.g. Ts 15 cm☼)}

\item{number_consec_days}{Minimum number of days used to define the growing season, for which air temperature, soil temperature or soil temperature amplitude remains above a set threshold}
}
\value{
a summarized dataframe, containing the aggregation timescale variables and the aggregated values
}
\description{
Biomet variables aggregator, from halfhourly to either hourly, daily, weekly, monthly yearly and growing_season timescales
}
