% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_montecarlo_sims.R
\name{merge_montecarlo_sims}
\alias{merge_montecarlo_sims}
\title{merge all gapfilled montecarlo simulations into a single file}
\usage{
merge_montecarlo_sims(
  gf_type = "rf",
  dir,
  saving_dir,
  datetime1 = "datetime",
  filled_flux_col1 = "ch4_flux_final_filled"
)
}
\arguments{
\item{gf_type}{the type of gapfilling performed on the montecarlo simulations. Possible values are only "rf" or "REddyProc", which determine how to read the files}

\item{dir}{folder containing only the individual files to be merged}

\item{saving_dir}{folder where the merged file will be saved}

\item{datetime1}{name of the timestamp column}

\item{filled_flux_col1}{name of the filled flux column}
}
\value{
merged dataframe, also saved on disk
}
\description{
merge all gapfilled montecarlo simulations into a single file
}
