% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddyproc_gapfiller.R
\name{reddyproc_gapfiller}
\alias{reddyproc_gapfiller}
\title{A wrapper for ReddyProc gapfiller for gapfilling NEE or H2O fluxes}
\usage{
reddyproc_gapfiller(
  formatted_file_path,
  saving_folder,
  FLUX = c("NEE", "H2O"),
  longitude = 19.556646,
  latitude = 64.18198,
  timezone = 1
)
}
\arguments{
\item{formatted_file_path}{path to the file ready for ReddyProc gapfilling. Typically the output from the "formatting_fluxes_REddyProc" function of the PostEddyPro package}

\item{saving_folder}{path to the folder where the gapfilled fluxes will be saved}

\item{FLUX}{flux column in the final output: NEE for CO2 flux OR H2O for water flux}

\item{longitude}{longitude of the site}

\item{latitude}{latitude of the site}

\item{timezone}{timezone as a signed integer (e.g. 1 for UTC+1)}
}
\value{
No value returned, but a file saved to disk
}
\description{
A wrapper for ReddyProc gapfiller for gapfilling NEE or H2O fluxes
}
