% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualization}
\alias{visualization}
\title{Merge Eddypro and Biomet and Visualize fluxes}
\usage{
visualization(path_EC, path_biomet, date_format = "\%Y-\%m-\%d")
}
\arguments{
\item{path_EC}{The path to your untouched Eddypro full output}

\item{path_biomet}{The path to your merged biomet file}

\item{date_format}{If the eddypro output has not been altered, the date format is "\%Y-\%m-\%d", the default. If it has been opened in excel for instance, the date format is likely "\%d/\%m/\%Y"}
}
\value{
a named list, $data is a dataframe merged from flux and biomet files, $graphs is a graph of all gas fluxes
}
\description{
Merge Eddypro and Biomet and Visualize fluxes
}
