% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_gapfiller.R
\name{xgboost_gapfiller}
\alias{xgboost_gapfiller}
\title{An XGBoost gapfiller, with an integrated tuning-training-testing from caret}
\usage{
xgboost_gapfiller(
  site_df,
  datetime = "datetime",
  flux_col = "co2_flux",
  preds,
  sitename,
  tuning_grid = list(nrounds = c(500, 1000, 1500), max_depth = c(5, 10, 15), eta =
    c(0.05, 0.3), gamma = c(0, 0.5), colsample_bytree = c(0.65, 0.75), min_child_weight =
    c(2, 5), subsample = c(0.5, 0.75))
)
}
\arguments{
\item{site_df}{dataframe containing all a POSIXct timestamp, the flux data and predictor columns (preds argument).}

\item{datetime}{name of the timestamp (datetime) column already in POSIXct format}

\item{flux_col}{flux column to be gapfilled, this function was developped for methane fluxes, but can be used for other fluxes as long as a good set of predictors are provided}

\item{preds}{vector of predictors that have been gapfilled in advance}

\item{sitename}{sitename just for labelling the plots}

\item{tuning_grid}{named list of hyperparameters to tune, nrounds, max_depth,
eta, gamma, colsample_bytree, min_child_weight, subsample. Read more about it here https://www.hackerearth.com/practice/machine-learning/machine-learning-algorithms/beginners-tutorial-on-xgboost-parameter-tuning-r/tutorial/}
}
\value{
large list of elements. tuningmodel: finalmodel (the model used for gapfilling), site_df (the final data.frame containing both the input dataframe and the gapfilled column), r_sq_cv (crossvalidated R-squared)
rmse_cv (crossvalidated rmse), pred_meas (graph showing predicted vs. measured values), gf_meas_time (a graph showing both measured and gapfilled values with time)
}
\description{
An XGBoost gapfiller, with an integrated tuning-training-testing from caret
}
